+++ src/evaluation/voiceAnalysis.py
@@ -140,7 +140,7 @@ def get_filterbanks(nfilt=20,nfft=512,samplerate=16000,lowfreq=0,highfreq=None):
     #  from Hz to fft bin number
     bin = numpy.floor((nfft+1)*mel2hz(melpoints)/samplerate)
 
-    fbank = numpy.zeros([nfilt,nfft/2+1])
+    fbank = numpy.zeros([nfilt,int(nfft/2+1)])
     for j in range(0,nfilt):
         for i in range(int(bin[j]),int(bin[j+1])):
             fbank[j,i] = (i - bin[j])/(bin[j+1]-bin[j])
